backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
faster-whisper==1.0.1
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
httpx==0.27.0
groq==0.5.0
